requirements.txt
aiofiles
dill
dnspython
database-connect
evidently
Flask
httptools
imblearn
ipykernel
jinja2
mypy-boto3-s3
pip-chill
pymongo
python-dotenv
python-multipart
seaborn
types-s3transfer
uvicorn
watchfiles
websockets
wincertstore
neuro-mf
boto3
python-dotenv
kneed
-e .